requirements.txt
streamlit
streamlit_sortables
gspread
google-auth
pandas
gspread_dataframe